removed const from some function prototypes and inline functions

+++ ow_materials/include/AxisAlignedGrid.tpp
@@ -48,7 +48,9 @@ AxisAlignedGrid<T>::AxisAlignedGrid(
   );
 
   // restrict all dimensions to cube root of the max size_t to prevent overflow
-  const auto max_dimension = std::cbrt(std::numeric_limits<size_t>::max());
+  const auto max_dimension = std::cbrt(
+    static_cast<float>(std::numeric_limits<size_t>::max())
+  );
   if (   m_dimensions.X() >= max_dimension
       || m_dimensions.Y() >= max_dimension
       || m_dimensions.Z() >= max_dimension) {
